tests/wpt/tests/css/css-tables/caption-writing-mode-002.html

<!DOCTYPE html>
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<link rel="author" title="David Grogan" href="dgrogan@chromium.org">
<link rel="help" href="https://www.w3.org/TR/css-writing-modes-3/#orthogonal-flows">
<style>
x-table {
  display: table;
  width: 300px;
  height: 300px;
  writing-mode: vertical-lr;
  outline: 2px dashed blue;
}
x-caption {
  display: table-caption;
  height: 200px;
  width: 200px;
  writing-mode: horizontal-tb;
  outline: 1px solid black;
}
</style>

<x-table>
  <x-caption>caption</x-caption>
</x-table>

<script>
let body_height = document.querySelector("body").offsetHeight;
test(() => {
  assert_greater_than(body_height, 0, "Vertical table with horizontal caption doesn't make <body> have crazy height.");
  assert_less_than(body_height, 10000, "Vertical table with horizontal caption doesn't make <body> have crazy height.");
})
</script>
